src/requirements.txt
aioredis==1.3.1
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
backoff==2.2.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.3
Deprecated==1.2.14
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-pagination==0.12.14
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
hiredis==2.3.2
idna==3.6
MarkupSafe==2.1.5
orjson==3.8.2
packaging==23.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.2
pydantic-settings==2.2.1
python-dotenv==0.21.0
python-jose==3.3.0
redis==4.3.4
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.15.0
uvloop==0.17.0
Werkzeug==3.0.1
wrapt==1.16.0
alembic==1.13.1
typer==0.9.0
async-typer==0.1.8
fastapi-jwt==0.2.0
aiohttp==3.9.3
httpx==0.27.0
pytest==8.0.2